% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggauto.R
\name{ggauto}
\alias{ggauto}
\title{Automatic ggplot plot}
\usage{
ggauto(
  var1 = NULL,
  var2 = NULL,
  var3 = NULL,
  data = NULL,
  xlab = "x",
  ylab = "y",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  base_size = 14,
  base_family = "sans"
)
}
\arguments{
\item{var1}{First variable. Either vector (for example \code{plot_data$v1}), or quoted character string with a column name (for example \code{"v1"}) when passing in data.}

\item{var2}{Optional second variable.}

\item{var3}{Optional third variable.}

\item{data}{Optional data frame to get variables from.}

\item{xlab}{Label for the x-axis. Defaults to \code{"x"}.}

\item{ylab}{Label for the y-axis. Defaults to \code{"y"}.}

\item{title}{Optional plot title.}

\item{subtitle}{Optional plot subtitle.}

\item{caption}{Optional plot caption.}

\item{base_size}{Base font size. Defaults to \code{14}.}

\item{base_family}{Base font family. Defaults to \code{"sans"}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Automatically create an appropriate ggplot2 chart type based on the
classes of the provided variables.
}
\examples{
ggauto(var1 = mtcars$mpg)

}
